// TheProject/types/preferenceType.ts
export interface IRegionPreference {
  Gangnam: boolean;
  Seolleung: boolean;
  Samsung: boolean;
  Yeoksam: boolean;
  Nonhyeon: boolean;
}

export interface IObjPreference {
  PersonalStudy: boolean;
  Laptop: boolean;
  Non_face_to_face: boolean;
  Face_to_face: boolean;
}

export interface IWorkspacePreference {
  Cafe: boolean;
  Studyroom: boolean;
  Studycafe: boolean;
  Meetingroom: boolean;
  SharedOffice: boolean;
  Library: boolean;
  Bookcafe: boolean;
  Lounge: boolean;
}